[Core] Split Time class definitions into a dedicated header
* Generated a separate Time module header and TDL definition for the Time class
* Moved shared time-zone structures into common core definitions
* Added Time header includes for modules that instantiate Time objects
