+++ docs/xml/modules/classes/time.xml
@@ -11,7 +11,7 @@
     <id>0904cf78</id>
     <idstring>TIME</idstring>
     <category>System</category>
-    <copyright>Paul Manias 1996-2026</copyright>
+    <copyright>Paul Manias © 1996-2026</copyright>
     <description>
 <p>The Time class is available for programs that require time and date management in a multi-platform manner.</p>
 <p>To get the current system time, use the <action>Query</action> action.</p></description>
@@ -166,19 +166,5 @@
 
   </fields>
   <structs>
-    <struct name="TimeZoneInfo" comment="Time zone information returned from GetTimeZoneInfo()">
-      <field name="ZoneID" type="STRING">Preferred public ID, e.g. Europe/London or UTC.</field>
-      <field name="NativeID" type="STRING">Host-native ID, e.g. GMT Standard Time on Windows.</field>
-      <field name="Source" type="STRING">"tzif", "win32", or "utc".</field>
-      <field name="DataPath" type="STRING">TZif path on Linux, otherwise empty.</field>
-      <field name="Version" type="STRING">TZDB/source version if available, otherwise empty.</field>
-      <field name="Transitions" type="kt::vector&lt;TimeZoneTransition&gt;">Transitions available for the requested range.</field>
-      <field name="BaseOffset" type="INT">Standard UTC offset in seconds.</field>
-      <field name="StartYear" type="INT16">Inclusive requested start year.</field>
-      <field name="EndYear" type="INT16">Inclusive requested end year.</field>
-      <field name="IsLocal" type="INT8">1 if ZoneID requested the local system zone.</field>
-      <field name="IsFallback" type="INT8">1 if UTC fallback was used.</field>
-    </struct>
-
   </structs>
 </book>
