+++ src/core/defs/core.tdl
@@ -10,7 +10,6 @@ module({ name="Core", copyright="Paul Manias 1996-2026", timestamp=20241015,
 
   privateNames({ "CoreBase", "DateTime", "ResourceManager" })
 
-  loadFile(glPath .. 'common.tdl')
   loadFile(glPath .. 'common-graphics.tdl')
 
   flags("MSF", { comment="Message flags." },
@@ -576,35 +575,6 @@ using ModTest    = void (*)(std::string_view, int *, int *);
    template <class T> FieldDef(CSTRING pName, T pValue) : Name(pName), Value(int(pValue)) { }
   ]])
 
-  struct("TimeZoneTransition", { comment="Time zone transition information for GetTimeZoneInfo()" }, [[
-     large Instant        # Unix epoch microseconds at which the transition occurs.
-     string Abbreviation  # Short name where available, e.g. GMT, BST, PST.
-     int OffsetBefore     # UTC offset in seconds before the transition.
-     int OffsetAfter      # UTC offset in seconds after the transition.
-     int DaylightSaving   # 1 if the resulting period observes daylight-saving time.
-  ]],
-  [[
-   TimeZoneTransition() : Instant(0), OffsetBefore(0), OffsetAfter(0), DaylightSaving(0) { }
-  ]])
-
-  struct("TimeZoneInfo", { comment="Time zone information returned from GetTimeZoneInfo()" }, [[
-     string ZoneID      # Preferred public ID, e.g. Europe/London or UTC.
-     string NativeID    # Host-native ID, e.g. GMT Standard Time on Windows.
-     string Source      # "tzif", "win32", or "utc".
-     string DataPath    # TZif path on Linux, otherwise empty.
-     string Version     # TZDB/source version if available, otherwise empty.
-     vector(struct(TimeZoneTransition)) Transitions # Transitions available for the requested range.
-     int BaseOffset     # Standard UTC offset in seconds.
-     short StartYear    # Inclusive requested start year.
-     short EndYear      # Inclusive requested end year.
-     char IsLocal       # 1 if ZoneID requested the local system zone.
-     char IsFallback    # 1 if UTC fallback was used.
-  ]],
-  [[
-   TimeZoneInfo() : BaseOffset(0), StartYear(0), EndYear(0), IsLocal(0), IsFallback(0) { }
-   TimeZoneInfo(int) : TimeZoneInfo() { }
-  ]])
-
   struct("SystemState", { comment="Returned by the GetSystemState() function." }, [[
     cstr Platform      # String-based field indicating the user's platform.  Currently returns `Native`, `Windows`, `OSX` or `Linux`.
     cstr IDL           # The Core module's compressed IDL string
@@ -1056,6 +1026,7 @@ typedef std::vector<obj_write> WRITE_TABLE;
   klass("StorageDevice")
   klass("Thread")
   klass("Task")
+  klass("Time")
 
   methods("MetaClass", "mc", {
     { id=1, name="FindField" }
@@ -1126,27 +1097,6 @@ inline CLASSID Object::baseClassID() { return Class->BaseClassID; }
       }
       else return ERR::CreateObject;
    }
-  ]])
-
-  -- Note that the fields following SystemTime are identical to struct DateTime
-
-  methods("Time", "Pt", {
-    { id=1, name="SetTime" },
-    { id=2, name="GetTimeZoneInfo" }
-  })
-
-  klass("Time", { src="../classes/class_time.cpp", output="../classes/class_time_def.c" }, [[
-    large SystemTime  # Total number of microseconds passed since the system base time
-    int Year          # Year   (-ve for BC, +ve for AD)
-    int Month         # Month  (1 - 12)
-    int Day           # Day    (1 - 31)
-    int Hour          # Hour   (0 - 23)
-    int Minute        # Minute (0 - 59)
-    int Second        # Second (0 - 59)
-    int TimeZone      # 0 is GMT, range is +1300 or -1300
-    int DayOfWeek     # Day of week (0 = Sunday ..)
-    int MilliSecond   # Millisecond (0 - 999)
-    int MicroSecond   # Microsecond  (0 - 999999)
   ]])
 
   c_insert([==[
