+++ src/core/defs/filesystem.tdl
@@ -5,7 +5,6 @@ header({ module="Core", copyright="Paul Manias © 1996-2026", timestamp=20240611
 
   restrict(function()
     loadFile('./common.tdl')
-    loadFile(glPath .. 'common.tdl')
   end)
 
   klass("StorageDevice", { src="../classes/class_storagedevice.cpp", output="../classes/class_storagedevice_def.cpp" }, [[
