+++ src/document/document.cpp
@@ -36,6 +36,7 @@ that is distributed with this package.  Please refer to it for further informati
 #include <kotuku/modules/processes.h>
 #include <kotuku/modules/script.h>
 #include <kotuku/modules/xquery.h>
+#include <kotuku/modules/time.h>
 #include <atomic>
 
 using BYTECODE = uint32_t;
